move write_records into lance client
